test fn calcTime
