+++ .eslintrc.json
@@ -9,7 +9,8 @@
         "prettier"
       ],
       "parserOptions": {
-        "ecmaVersion": 2020
+        "sourceType": "module",
+        "ecmaVersion": 6
       },
       "rules": {
         "prettier/prettier": "error",
