+++ index.html
@@ -5,6 +5,7 @@
         <meta http-equiv="X-UA-Compatible" content="IE=edge" />
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
         <title>Page 1</title>
+        
     </head>
     <body>
         <section class="main">
@@ -57,7 +58,6 @@
                 <h3 class="popup__head">Вы выиграли!</h3>
                 <h3 class="popup__title">Затраченное время:</h3>
                 <p class="popup__title popup__title--time"></p>
-                <div class="popup__table"></div>
                 <button class="popup__btn popup__btn--game" type="submit">
                     Играть снова
                 </button>
@@ -70,11 +70,11 @@
                 <h3 class="popup__head">Вы проиграли!</h3>
                 <h3 class="popup__title">Затраченное время:</h3>
                 <p class="popup__title popup__title--time"></p>
-                <div class="popup__table"></div>
                 <button class="popup__btn popup__btn--game" type="submit">
                     Играть снова
                 </button>
             </div>
         </section>
+        
     </body>
 </html>
