+++ package.json
@@ -12,29 +12,32 @@
         "lint": "npm run lint:js && npm run lint:css",
         "build": "cross-env NODE_ENV=production webpack",
         "start": "webpack-dev-server --open"
+        
     },
     "author": "",
     "license": "ISC",
     "devDependencies": {
-        "copy-webpack-plugin": "^11.0.0",
-        "cross-env": "^7.0.3",
-        "css-loader": "^6.7.3",
-        "css-minimizer-webpack-plugin": "^5.0.0",
-        "eslint": "^8.38.0",
-        "eslint-config-prettier": "^8.8.0",
-        "eslint-plugin-prettier": "^4.2.1",
-        "html-webpack-plugin": "^5.5.1",
-        "mini-css-extract-plugin": "^2.7.5",
-        "prettier": "^2.8.7",
-        "style-loader": "^3.3.2",
-        "stylelint": "^14.16.1",
-        "stylelint-config-prettier": "^9.0.4",
-        "stylelint-config-standard": "^29.0.0",
-        "stylelint-prettier": "^2.0.0",
-        "ts-loader": "^9.4.2",
-        "typescript": "^5.0.4",
-        "webpack": "^5.81.0",
-        "webpack-cli": "^5.0.2",
-        "webpack-dev-server": "^4.13.3"
+        "@testing-library/dom": "9.2.0",
+        "copy-webpack-plugin": "11.0.0",
+        "cross-env": "7.0.3",
+        "css-loader": "6.7.3",
+        "css-minimizer-webpack-plugin": "5.0.0",
+        "eslint": "8.38.0",
+        "eslint-config-prettier": "8.8.0",
+        "eslint-plugin-prettier": "4.2.1",
+        "html-webpack-plugin": "5.5.1",
+        "jest": "29.5.0",
+        "mini-css-extract-plugin": "2.7.5",
+        "prettier": "2.8.7",
+        "style-loader": "3.3.2",
+        "stylelint": "14.16.1",
+        "stylelint-config-prettier": "9.0.4",
+        "stylelint-config-standard": "29.0.0",
+        "stylelint-prettier": "2.0.0",
+        "ts-loader": "9.4.2",
+        "typescript": "5.0.4",
+        "webpack": "5.81.0",
+        "webpack-cli": "5.0.2",
+        "webpack-dev-server": "4.13.3"
     }
 }
