+++ style.css
@@ -2,6 +2,7 @@
     margin: 0;
     padding: 0;
 }
+
 .main {
     width: 100%;
     height: 100vh;
@@ -10,6 +11,7 @@
     justify-content: center;
     align-items: center;
 }
+
 .blue_block {
     width: 480px;
     height: 459px;
@@ -25,10 +27,12 @@
     src: local('StratosSkyeng Medium'),
         url(./fonts/statosskyeng-v1.001/Desktop\ fonts/stratosskyeng-medium.otf);
 }
+
 .blue_block_head {
     width: 208px;
     height: 96px;
 }
+
 .blue_block_head_h1 {
     font-family: 'StratosSkyeng Medium';
     font-style: normal;
@@ -39,11 +43,13 @@
     font-feature-settings: 'pnum' on, 'lnum' on;
     color: #004980;
 }
+
 .blue_block_items {
     display: flex;
     gap: 25px;
     margin-top: 48px;
 }
+
 .blue_block_items_all {
     width: 98px;
     height: 98px;
@@ -59,6 +65,7 @@
     color: #0080c1;
     justify-content: center;
 }
+
 .button {
     width: 246px;
     height: 50px;
@@ -74,6 +81,7 @@
     font-feature-settings: 'pnum' on, 'lnum' on;
     color: #ffffff;
 }
+
 .button:hover {
     background: #0080c1;
 }
@@ -174,11 +182,10 @@
 
 .cards__items {
     padding-top: 48px;
-    display: flex;
-    flex-wrap: wrap;
-    row-gap: 15px;
-    column-gap: 15px;
-    justify-content: space-around;
+    display: grid;
+    grid-template-columns: repeat(6, 1fr);
+    row-gap: 35px;
+    justify-items: center;
 }
 
 .cards__item--turned {
@@ -199,13 +206,7 @@
     }
 }
 
-.cards__item--low-difficulty {
-    width: 10%;
-    height: 10%;
-}
-
 .popup {
-    /* в скрытом состоянии */
     display: none;
     position: fixed;
     width: 460px;
@@ -215,7 +216,6 @@
 }
 
 .popup--show {
-    /* Отображет popup */
     display: block;
 }
 
@@ -227,6 +227,7 @@
     background: #c2f5ff;
     border-radius: 12px;
     text-align: center;
+    padding-top: 20px;
 }
 
 .popup__head {
