--- tsconfig.json
@@ -1,12 +1,0 @@
-{
-    "compilerOptions": {
-      "strict": true, // Включает все жесткие проверки типов
-      "outDir": "./dist/",
-      "noImplicitAny": false,
-      "module": "es6",
-      "target": "es5",
-      "allowJs": true,
-      "moduleResolution": "node",
-      "strictNullChecks": false
-    }
-  }
